NEX-1661 Tab bars with icons for activity and treasury
